Fix SearchVC
