+++ YouTube/Classes/CustomViews/HomeCell/HomeCell.xib
@@ -67,7 +67,7 @@
                                 <color key="highlightedColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
                             </label>
                         </subviews>
-                        <color key="backgroundColor" cocoaTouchSystemColor="darkTextColor"/>
+                        <color key="backgroundColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
                         <constraints>
                             <constraint firstItem="uwS-mR-7Dg" firstAttribute="leading" secondItem="aZY-hZ-5L7" secondAttribute="leading" constant="10" id="8pD-KH-mef"/>
                             <constraint firstItem="V1t-n3-4tY" firstAttribute="leading" secondItem="KFH-5p-tKy" secondAttribute="trailing" constant="8" id="Ceb-Or-RB4"/>
