+++ YouTube/Classes/CustomViews/ResultSearchCell/ResultSearchCell.xib
@@ -8,7 +8,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="36" id="KGk-i7-Jjw" customClass="ResultSearchCell" customModule="YouTube" customModuleProvider="target">
+        <tableViewCell alpha="0.80000000000000004" contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="36" id="KGk-i7-Jjw" customClass="ResultSearchCell" customModule="YouTube" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="320" height="36"/>
             <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
             <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
@@ -16,21 +16,22 @@
                 <autoresizingMask key="autoresizingMask"/>
                 <subviews>
                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Y3f-V6-Lup">
-                        <rect key="frame" x="16" y="8" width="288" height="20"/>
+                        <rect key="frame" x="54" y="8" width="250" height="20"/>
                         <constraints>
                             <constraint firstAttribute="height" constant="20" id="qdd-wv-Ngd"/>
                         </constraints>
                         <fontDescription key="fontDescription" name="HelveticaNeue" family="Helvetica Neue" pointSize="15"/>
-                        <color key="textColor" cocoaTouchSystemColor="darkTextColor"/>
+                        <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
                         <nil key="highlightedColor"/>
                     </label>
                 </subviews>
                 <constraints>
                     <constraint firstAttribute="trailingMargin" secondItem="Y3f-V6-Lup" secondAttribute="trailing" constant="8" id="Tp1-nM-oh3"/>
-                    <constraint firstItem="Y3f-V6-Lup" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leadingMargin" constant="8" id="u6o-hh-zqI"/>
+                    <constraint firstItem="Y3f-V6-Lup" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leadingMargin" constant="46" id="u6o-hh-zqI"/>
                     <constraint firstItem="Y3f-V6-Lup" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="topMargin" id="udj-Md-aBm"/>
                 </constraints>
             </tableViewCellContentView>
+            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="calibratedWhite"/>
             <connections>
                 <outlet property="nameVideoLabel" destination="Y3f-V6-Lup" id="szS-Yl-vgo"/>
             </connections>
