--- YouTube/Classes/ViewControllers/VideoSearchViewController/VideoSearchViewController.xib
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="10117" systemVersion="15F34" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES">
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="10085"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="VideoSearchViewController" customModule="YouTube" customModuleProvider="target">
-            <connections>
-                <outlet property="videoSearchTableView" destination="EQe-GA-aei" id="EHT-CT-bZE"/>
-                <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
-            <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="zIY-ub-rVa">
-                    <rect key="frame" x="0.0" y="0.0" width="375" height="64"/>
-                    <subviews>
-                        <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Jft-96-CRF">
-                            <rect key="frame" x="8" y="24" width="32" height="32"/>
-                            <state key="normal" backgroundImage="bt_back"/>
-                            <connections>
-                                <action selector="backToSearchViewController:" destination="-1" eventType="touchUpInside" id="vSF-G7-s8T"/>
-                            </connections>
-                        </button>
-                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Video" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="FhU-uE-LbJ">
-                            <rect key="frame" x="155.5" y="26.5" width="63" height="29.5"/>
-                            <fontDescription key="fontDescription" name="HelveticaNeue" family="Helvetica Neue" pointSize="25"/>
-                            <color key="textColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                            <nil key="highlightedColor"/>
-                        </label>
-                    </subviews>
-                    <color key="backgroundColor" red="1" green="0.1450980392" blue="0.08235294118" alpha="1" colorSpace="calibratedRGB"/>
-                    <constraints>
-                        <constraint firstItem="Jft-96-CRF" firstAttribute="leading" secondItem="zIY-ub-rVa" secondAttribute="leading" constant="8" id="Cdc-hV-Gp1"/>
-                        <constraint firstAttribute="bottom" secondItem="Jft-96-CRF" secondAttribute="bottom" constant="8" id="Jeh-FR-6nn"/>
-                        <constraint firstAttribute="bottom" secondItem="FhU-uE-LbJ" secondAttribute="bottom" constant="8" id="SK2-oC-9wH"/>
-                        <constraint firstItem="FhU-uE-LbJ" firstAttribute="centerX" secondItem="zIY-ub-rVa" secondAttribute="centerX" id="W4Q-To-1mP"/>
-                        <constraint firstAttribute="height" constant="64" id="epi-kM-nOp"/>
-                    </constraints>
-                </view>
-                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="none" rowHeight="44" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="EQe-GA-aei">
-                    <rect key="frame" x="0.0" y="64" width="375" height="563"/>
-                    <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                    <connections>
-                        <outlet property="dataSource" destination="-1" id="dw0-M3-YTA"/>
-                        <outlet property="delegate" destination="-1" id="Ja5-If-pHs"/>
-                    </connections>
-                </tableView>
-            </subviews>
-            <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="calibratedWhite"/>
-            <constraints>
-                <constraint firstItem="zIY-ub-rVa" firstAttribute="top" secondItem="i5M-Pr-FkT" secondAttribute="top" id="6U9-uv-Tmd"/>
-                <constraint firstItem="EQe-GA-aei" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" id="FZG-V0-tDz"/>
-                <constraint firstAttribute="bottom" secondItem="EQe-GA-aei" secondAttribute="bottom" constant="40" id="JyK-ji-TMj"/>
-                <constraint firstAttribute="trailing" secondItem="EQe-GA-aei" secondAttribute="trailing" id="Xb9-Pl-B56"/>
-                <constraint firstItem="EQe-GA-aei" firstAttribute="top" secondItem="zIY-ub-rVa" secondAttribute="bottom" id="qVd-SZ-Gs3"/>
-                <constraint firstAttribute="trailing" secondItem="zIY-ub-rVa" secondAttribute="trailing" id="ssg-A4-Ps1"/>
-                <constraint firstItem="zIY-ub-rVa" firstAttribute="leading" secondItem="i5M-Pr-FkT" secondAttribute="leading" id="usl-6G-Enl"/>
-            </constraints>
-            <simulatedScreenMetrics key="simulatedDestinationMetrics" type="retina47"/>
-            <point key="canvasLocation" x="224.5" y="195.5"/>
-        </view>
-    </objects>
-    <resources>
-        <image name="bt_back" width="32" height="32"/>
-    </resources>
-</document>
